% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{methods}
\name{dbDisconnect,ODBCConnection-method}
\alias{dbDisconnect,ODBCConnection-method}
\title{Close a current session.}
\usage{
\S4method{dbDisconnect}{ODBCConnection}(conn)
}
\arguments{
\item{conn}{a \code{\linkS4class{ODBCConnection}} object, produced by \code{\link[DBI]{dbConnect}}}
}
\description{
Close a current session.
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(RODBCDBI::ODBC(), dsn="test", user="sa", password="Password12!")
dbDisconnect(con)
}
}

