% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODBCResult.R
\name{dbDone}
\alias{dbDone}
\title{Hack its is_done property assigned in the parent environment.}
\usage{
dbDone(res, n, boolean)
}
\description{
Hack its is_done property assigned in the parent environment.
}

