% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{methods}
\name{dbExistsTable,ODBCConnection,character-method}
\alias{dbExistsTable,ODBCConnection,character-method}
\title{Does the table exist?}
\usage{
\S4method{dbExistsTable}{ODBCConnection,character}(conn, name)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{ODBCConnection}}}

\item{name}{String, name of table. Match is case insensitive.}
}
\description{
Does the table exist?
}

