% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{methods}
\name{dbGetInfo,ODBCConnection-method}
\alias{dbGetInfo,ODBCConnection-method}
\title{Get DBMS metadata.}
\usage{
\S4method{dbGetInfo}{ODBCConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from \code{\linkS4class{ODBCConnection}}, \code{\linkS4class{ODBCDriver}}, or a \code{\linkS4class{ODBCResult}}}
}
\description{
Get DBMS metadata.
}

