% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{methods}
\name{dbListFields,ODBCConnection,character-method}
\alias{dbListFields,ODBCConnection,character-method}
\title{List fields in specified table.}
\usage{
\S4method{dbListFields}{ODBCConnection,character}(conn, name)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{RODBCConnection}}}

\item{name}{a length 1 character vector giving the name of a table.}
}
\description{
List fields in specified table.
}
\examples{
library(DBI)
con <- dbConnect(RODBCDBI::ODBC(), dsn="test", user="sa", password="Password12!")
dbWriteTable(con, "iris", iris, overwrite=TRUE)
dbListFields(con, "iris")
dbDisconnect(con)
}

