% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{methods}
\name{dbListTables,ODBCConnection-method}
\alias{dbListTables,ODBCConnection-method}
\title{List available ODBC tables.}
\usage{
\S4method{dbListTables}{ODBCConnection}(conn)
}
\arguments{
\item{conn}{An existing \code{\linkS4class{ODBCConnection}}}
}
\description{
List available ODBC tables.
}

