% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{methods}
\name{dbWriteTable,ODBCConnection,character,data.frame-method}
\alias{dbWriteTable,ODBCConnection,character,data.frame-method}
\title{Write a local data frame or file to the database.}
\usage{
\S4method{dbWriteTable}{ODBCConnection,character,data.frame}(conn, name, value,
  overwrite = FALSE, append = FALSE, ...)
}
\arguments{
\item{name}{a character string specifying a table name. ODBCConnection table names
are \emph{not} case sensitive, e.g., table names \code{ABC} and \code{abc}
are considered equal.}

\item{value}{a data.frame (or coercible to data.frame) object or a
file name (character).  when \code{value} is a character, it is interpreted as a file name and its contents imported to ODBC.}

\item{con,conn}{a \code{\linkS4class{ODBCConnection}} object, produced by \code{\link[DBI]{dbConnect}}}
}
\description{
Write a local data frame or file to the database.
}
\examples{
library(DBI)
con <- dbConnect(RODBCDBI::ODBC(), dsn="test", user="sa", password="Password12!")
dbWriteTable(con, "mtcars", mtcars, overwrite=TRUE)
dbReadTable(con, "mtcars")
dbDisconnect(con)
}

