% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{class}
\name{ODBCConnection-class}
\alias{ODBCConnection-class}
\title{Class ODBCConnection.}
\description{
\code{ODBCConnection} objects are usually created by \code{\link[DBI]{dbConnect}}
}
\keyword{internal}

