% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCDriver.R
\docType{class}
\name{ODBCDriver-class}
\alias{ODBCDriver-class}
\alias{dbGetInfo,ODBCDriver-method}
\alias{dbIsValid,ODBCDriver-method}
\alias{dbUnloadDriver,ODBCDriver-method}
\title{ODBCDriver and methods.}
\usage{
\S4method{dbUnloadDriver}{ODBCDriver}(drv, ...)

\S4method{dbIsValid}{ODBCDriver}(dbObj)

\S4method{dbGetInfo}{ODBCDriver}(dbObj, ...)
}
\description{
An ODBC driver implementing the R database (DBI) API.
This class should always be initialized with the \code{ODBC()} function.
It returns an object that allows you to connect to ODBC.

Nothing to do for ODBCDriver case
}
\keyword{internal}

