% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCResult.R
\docType{class}
\name{ODBCResult-class}
\alias{ODBCResult-class}
\title{Class ODBCResult.}
\description{
ODBC's query results class. This classes encapsulates the result of an SQL statement (either select or not). The main generator is dbSendQuery.
}
\keyword{internal}

