% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RODBCDBI.R
\docType{package}
\name{RODBCDBI}
\alias{RODBCDBI}
\alias{RODBCDBI-package}
\title{RODBCDBI}
\description{
Provides Access to Databases Through the ODBC Interface An implementation of R's DBI interface using ODBC package as a back-end. This allows R to connect to any DBMS that has a ODBC driver.
}

