% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODBCResult.R
\name{set_as_done}
\alias{set_as_done}
\title{Hack its is_done property assigned in the parent environment.}
\usage{
set_as_done(res, n, boolean)
}
\description{
Hack its is_done property assigned in the parent environment.
}

