% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODBCResult.R
\docType{methods}
\name{odbc-meta}
\alias{dbGetRowCount,ODBCResult-method}
\alias{dbGetStatement,ODBCResult-method}
\alias{odbc-meta}
\title{Database interface meta-data.}
\usage{
\S4method{dbGetRowCount}{ODBCResult}(res, ...)

\S4method{dbGetStatement}{ODBCResult}(res, ...)
}
\arguments{
\item{res}{An object of class \code{\linkS4class{ODBCResult}}}

\item{...}{Ignored. Needed for compatibility with generic}
}
\description{
See documentation of generics for more details.
}
\examples{
data(USArrests)
con <- dbConnect(SQLite(), dbname=":memory:")
dbWriteTable(con, "t1", USArrests)
dbWriteTable(con, "t2", USArrests)

dbListTables(con)

rs <- dbSendQuery(con, "select * from t1 where UrbanPop >= 80")
dbGetStatement(rs)
dbHasCompleted(rs)

info <- dbGetInfo(rs)
names(info)
info$fields

dbFetch(rs, n=2)
dbHasCompleted(rs)
info <- dbGetInfo(rs)
info$fields
dbClearResult(rs)

# DBIConnection info
names(dbGetInfo(con))

dbDisconnect(con)
}

