% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODBCConnection.R, R/ODBCResult.R
\docType{methods}
\name{dbSendQuery,ODBCConnection-method}
\alias{dbClearResult,ODBCResult-method}
\alias{dbFetch,ODBCResult-method}
\alias{dbHasCompleted,ODBCResult-method}
\alias{dbSendQuery,ODBCConnection-method}
\title{Execute a SQL statement on a database connection}
\usage{
\S4method{dbSendQuery}{ODBCConnection}(conn, statement, ...)

\S4method{dbFetch}{ODBCResult}(res, n = -1, ..., row.names = NA)

\S4method{dbHasCompleted}{ODBCResult}(res, ...)

\S4method{dbClearResult}{ODBCResult}(res, ...)
}
\arguments{
\item{res}{Code a \linkS4class{ODBCResult} produced by \code{\link[DBI]{dbSendQuery}}.}

\item{n}{Number of rows to return. If less than zero returns all rows.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

  If \code{TRUE}, always translate row names to a column called "row_names".
  If \code{FALSE}, never translate row names. If \code{NA}, translate
  rownames only if they're a character vector.

  A string is equivalent to \code{TRUE}, but allows you to override the
  default name.

  For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}
}
\description{
To retrieve results a chunk at a time, use \code{dbSendQuery},
\code{dbFetch}, then \code{ClearResult}. Alternatively, if you want all the
results (and they'll fit in memory) use \code{dbGetQuery} which sends,
fetches and clears for you.
}

