% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ODBCDriver.R
\name{ODBC}
\alias{ODBC}
\title{ODBC driver}
\usage{
ODBC()
}
\description{
This driver never needs to be unloaded and hence \code{dbUnload()} is a null-op.
}
\examples{
library(DBI)
RODBCDBI::ODBC()
}

