% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{class}
\name{ODBCConnection-class}
\alias{ODBCConnection-class}
\alias{dbGetInfo,ODBCConnection-method}
\title{Class ODBCConnection.}
\usage{
\S4method{dbGetInfo}{ODBCConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from \code{\linkS4class{ODBCConnection}}, \code{\linkS4class{ODBCDriver}}, or a \code{\linkS4class{ODBCResult}}}
}
\description{
\code{ODBCConnection} objects are usually created by \code{\link[DBI]{dbConnect}}

Get DBMS metadata.
}
\keyword{internal}

