% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODBCConnection.R
\docType{class}
\name{ODBCConnection-class}
\alias{ODBCConnection-class}
\alias{dbGetInfo,ODBCConnection-method}
\title{Class ODBCConnection.}
\usage{
\S4method{dbGetInfo}{ODBCConnection}(dbObj, ...)
}
\description{
\code{ODBCConnection} objects are usually created by \code{\link[DBI]{dbConnect}}
}
\keyword{internal}

