% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ODBCDriver.R
\docType{class}
\name{ODBCDriver-class}
\alias{ODBCDriver-class}
\alias{dbUnloadDriver,ODBCDriver-method}
\title{ODBCDriver and methods.}
\usage{
\S4method{dbUnloadDriver}{ODBCDriver}(drv, ...)
}
\description{
ODBCDriver and methods.
}
\keyword{internal}

